fix: replace toast.util.ts with toast.util.tsx to resolve conflict with main
