--- src/utils/toast.util.ts
@@ -1,22 +1,0 @@
-import toast from "react-hot-toast";
-
-const showToast = {
-	message: (message: string) => {
-		toast.remove();
-		toast(message);
-	},
-	success: (message: string) => {
-		toast.remove();
-		toast.success(message);
-	},
-	error: (message: string) => {
-		toast.remove();
-		toast.error(message);
-	},
-	loading: (message: string) => {
-		toast.remove();
-		toast.loading(message);
-	},
-};
-
-export default showToast;
